catalog component
return to prev page on click on browser back button
