+++ src/components/UserAccount/UserOrderHistoryDetailsList/UserOrderHistoryDetailsItem/UserOrderHistoryDetailsItem.jsx
@@ -1,4 +1,4 @@
-import { formatPrice } from "../../../utils/formatPrice";
+import { formatPrice } from "../../../../utils/formatPrice";
 import { Box, Image, InfoContainer, Item, Price, Quantity, Size } from "./UserOrderHistoryDetailsItem.styled";
 
 const UserOrderHistoryDetailsItem = ({ product }) => {
